feat: update usePermissions hook to request the permission to post notifications
